+++ examples/todomvc/src/App.tsx
@@ -10,8 +10,8 @@ import type { ActionFunctionArgs, LoaderFunctionArgs } from 'react-router-dom'
 import AppHeader from './components/AppHeader/index.jsx'
 import InputArea from './components/InputArea'
 import { W3APIProvider } from './components/W3API'
-import { List } from './List'
-import { AllLists } from './AllLists'
+import { List } from './components/List'
+import { AllLists } from './components/AllLists'
 import { LayoutProps, ListLoaderData, ListDoc } from './interfaces'
 import loadFixtures from './loadFixtures'
 import { useUploader, UploaderCtx } from './hooks/useUploader'
